adding better ci and build scripts
