+++ package.json
@@ -44,12 +44,11 @@
 	},
 	"scripts": {
 		"build": "node ./ci/tools/build-dist.js",
-		"lint": "eslint -c ./ci/.eslintrc.json ./src/**/*.js",
+		"lint": "eslint -c ./ci/.eslintrc.json ./ci/**/*.js ./src/**/*.js ./tests/**/*.js",
 		"test": "node ./node_modules/just-test/dist/tests-runner/run-tests.js --config=./tests/tests-run-config.json"
 	},
 	"devDependencies": {
 		"eslint": "^7.20.0",
-		"fs-extra": "^9.1.0",
 		"just-test": "2.3.2",
 		"uglify-es": "^3.3.9"
 	}
